Bench: 1533769

+++ src/psqt_state.hpp
@@ -31,13 +31,13 @@ public:
             auto& pieces                          = pos.piece_list(c);
             auto& squares                         = pos.piece_list_sq(c);
 
-            for (size_t i = 0; i < 16; i++) {
-                PieceType pt = pieces[static_cast<u8>(i)];
+            for (u8 i = 0; i < 16; i++) {
+                PieceType pt = pieces[i];
                 if (pt == PieceType::None) {
                     continue;
                 }
 
-                add_piece(c, pt, squares[static_cast<u8>(i)]);
+                add_piece(c, pt, squares[i]);
             }
         }
     }
